fix: arrumando classe usuario

+++ backend/devhub.api/src/main/java/com/devhub/api/domain/usuario/Usuario.java
@@ -17,7 +17,7 @@ import lombok.NoArgsConstructor;
 public abstract class Usuario {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    protected int id;
+    protected Long id;
 
     protected String nome;
